requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
openai>=1.3.0
twilio>=8.5.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.7.0
pydantic>=2.4.0
python-multipart>=0.0.6
aiofiles>=23.2.0
requests>=2.31.0
asyncio>=3.4.3
python-jose>=3.3.0